#494 UI Language noticeable , review changes
