+++ public/locales/en/translation.json
@@ -190,5 +190,5 @@
   "clipBoardCopied": "Copied",
   "clipBoardCopiedFailed": "Copy to clipboard failed",
   "landingHelpBtn": "Help",
-  "selectMultiLanguage": "Select Multi Language"
+  "selectMultiLanguage": "Select a Language"
 }
